Refactoring of code and comments
