+++ Raw_Data/Raw_Data.userprefs
@@ -7,7 +7,8 @@
   <MonoDevelop.Ide.ItemProperties.Raw__Data PreferredExecutionTarget="MonoDevelop.Default" />
   <MonoDevelop.Ide.Workbench ActiveDocument="Raw_Data\Program.cs">
     <Files>
-      <File FileName="Raw_Data\Program.cs" Line="1" Column="1" />
+      <File FileName="Raw_Data\Program.cs" Line="284" Column="4" />
+      <File FileName="Raw_Data\EmptyClass.cs" Line="1" Column="1" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MultiItemStartupConfigurations />
