--- Raw_Data/Raw_Data/Raw_Data.csproj.user
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition=" '$(RunConfiguration)' == 'Default' ">
-    <StartAction>Project</StartAction>
-  </PropertyGroup>
-</Project>
